+++ examples/CardRecurring/Controller.php
@@ -1,9 +1,9 @@
 <?php
 /**
- * This is the return controller for the EPS example.
- * It is called when the client is redirected back to the shop from the external page.
+ * This is the controller for the Card recurring example.
+ * It is called when the pay button on the index page is clicked.
  *
- * Copyright (C) 2018 heidelpay GmbH
+ * Copyright (C) 2019 heidelpay GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,11 +28,15 @@
 require_once __DIR__ . '/Constants.php';
 
 /** Require the composer autoloader file */
+/** @noinspection PhpIncludeInspection */
 require_once __DIR__ . '/../../../../autoload.php';
 
 use heidelpayPHP\examples\ExampleDebugHandler;
 use heidelpayPHP\Exceptions\HeidelpayApiException;
 use heidelpayPHP\Heidelpay;
+
+session_start();
+session_unset();
 
 $clientMessage = 'Something went wrong. Please try again later.';
 $merchantMessage = 'Something went wrong. Please try again later.';
@@ -45,13 +49,11 @@ function redirect($url, $merchantMessage = '', $clientMessage = '')
     die();
 }
 
-session_start();
-
-// Retrieve the paymentId you remembered within the Controller
-if (!isset($_SESSION['PaymentId'])) {
-    redirect(FAILURE_URL, 'The payment id is missing.', $clientMessage);
+// You will need the id of the payment type created in the frontend (index.php)
+if (!isset($_POST['resourceId'])) {
+    redirect(FAILURE_URL, 'Resource id is missing!', $clientMessage);
 }
-$paymentId = $_SESSION['PaymentId'];
+$paymentTypeId   = $_POST['resourceId'];
 
 // Catch API errors, write the message to your log and show the ClientMessage to the client.
 try {
@@ -59,15 +61,21 @@ try {
     $heidelpay = new Heidelpay(HEIDELPAY_PHP_PAYMENT_API_PRIVATE_KEY);
     $heidelpay->setDebugMode(true)->setDebugHandler(new ExampleDebugHandler());
 
-    // Redirect to success if the payment has been successfully completed.
-    $payment   = $heidelpay->fetchPayment($paymentId);
-    if ($payment->isCompleted()) {
+    $recurring = $heidelpay->activateRecurringPayment($paymentTypeId, RETURN_CONTROLLER_URL);
+
+    // You'll need to remember the paymentId for later in the ReturnController (in case of 3ds)
+    $_SESSION['PaymentTypeId'] = $paymentTypeId;
+    $_SESSION['ShortId'] = $recurring->getShortId();
+
+    // Redirect to the 3ds page or to success depending on the state of the transaction
+    if (empty($recurring->getRedirectUrl()) && $recurring->isSuccess()) {
         redirect(SUCCESS_URL);
+    } elseif (!empty($recurring->getRedirectUrl()) && $recurring->isPending()) {
+        redirect($recurring->getRedirectUrl());
     }
 
-    // Check the result message of the charge to find out what went wrong.
-    $charge = $payment->getChargeByIndex(0);
-    $merchantMessage = $charge->getMessage()->getCustomer();
+    // Check the result message of the transaction to find out what went wrong.
+    $merchantMessage = $recurring->getMessage()->getCustomer();
 } catch (HeidelpayApiException $e) {
     $merchantMessage = $e->getMerchantMessage();
     $clientMessage = $e->getClientMessage();
@@ -75,4 +83,3 @@ try {
     $merchantMessage = $e->getMessage();
 }
 redirect(FAILURE_URL, $merchantMessage, $clientMessage);
-