+++ examples/PayPalRecurring/ReturnController.php
@@ -1,9 +1,9 @@
 <?php
 /**
- * This is the return controller for the EPS example.
+ * This is the return controller for the PayPal example.
  * It is called when the client is redirected back to the shop from the external page.
  *
- * Copyright (C) 2018 heidelpay GmbH
+ * Copyright (C) 2019 heidelpay GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,12 +27,16 @@
 /** Require the constants of this example */
 require_once __DIR__ . '/Constants.php';
 
+/** @noinspection PhpIncludeInspection */
 /** Require the composer autoloader file */
 require_once __DIR__ . '/../../../../autoload.php';
 
 use heidelpayPHP\examples\ExampleDebugHandler;
 use heidelpayPHP\Exceptions\HeidelpayApiException;
 use heidelpayPHP\Heidelpay;
+use heidelpayPHP\Resources\PaymentTypes\Paypal;
+
+session_start();
 
 $clientMessage = 'Something went wrong. Please try again later.';
 $merchantMessage = 'Something went wrong. Please try again later.';
@@ -45,13 +49,11 @@ function redirect($url, $merchantMessage = '', $clientMessage = '')
     die();
 }
 
-session_start();
-
 // Retrieve the paymentId you remembered within the Controller
-if (!isset($_SESSION['PaymentId'])) {
-    redirect(FAILURE_URL, 'The payment id is missing.', $clientMessage);
+if (!isset($_SESSION['PaymentTypeId'])) {
+    redirect(FAILURE_URL, 'The payment type id is missing.', $clientMessage);
 }
-$paymentId = $_SESSION['PaymentId'];
+$paymentTypeId = $_SESSION['PaymentTypeId'];
 
 // Catch API errors, write the message to your log and show the ClientMessage to the client.
 try {
@@ -59,20 +61,17 @@ try {
     $heidelpay = new Heidelpay(HEIDELPAY_PHP_PAYMENT_API_PRIVATE_KEY);
     $heidelpay->setDebugMode(true)->setDebugHandler(new ExampleDebugHandler());
 
-    // Redirect to success if the payment has been successfully completed.
-    $payment   = $heidelpay->fetchPayment($paymentId);
-    if ($payment->isCompleted()) {
+    // Redirect to success if the payment has been successfully completed or is still in handled.
+    $paymentType = $heidelpay->fetchPaymentType($paymentTypeId);
+
+    if ($paymentType instanceof Paypal && $paymentType->isRecurring()) {
         redirect(SUCCESS_URL);
     }
-
-    // Check the result message of the charge to find out what went wrong.
-    $charge = $payment->getChargeByIndex(0);
-    $merchantMessage = $charge->getMessage()->getCustomer();
 } catch (HeidelpayApiException $e) {
     $merchantMessage = $e->getMerchantMessage();
     $clientMessage = $e->getClientMessage();
 } catch (RuntimeException $e) {
     $merchantMessage = $e->getMessage();
 }
-redirect(FAILURE_URL, $merchantMessage, $clientMessage);
 
+redirect(FAILURE_URL, $merchantMessage, $clientMessage);