+++ CHANGELOG.md
@@ -5,16 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [1.1.4.0][1.1.4.0]
 
-### Fixed
-*   Fixed existing examples.
-
 ### Added
 *   Enabled switching MGW environment (Prod, Staging, Dev) via environment variable.
 *   Examples for iDeal and Sofort.
 *   Added PHP 7.3 support.
+*   Add recurring payment for PayPal and Card payment type incl. examples.
 
 ### Changed
 *   Updated unit tests.
+*   Refactored examples.
 
 ## [1.1.3.0][1.1.3.0]